Add simple account creation example, simplify captcha example; not use WTForms
