+++ templates/create_account_form.html
@@ -1,40 +1,45 @@
 <!DOCTYPE html>
   <title>MediaWiki Create Account</title>
-  <!-- Bootstrap latest compiled and minified CSS -->
-  <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+
+  <link rel="stylesheet" href="static/bootstrap/css/bootstrap.min.css">
   <link rel="stylesheet" href="static/css/account_form.css">
-  <!-- Bootstrap optional theme -->
-  <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" charset="utf-8">
+
 <div class="container">
   <h2>Create MediaWiki Account</h2>
-  <form action="" method="post" role="form">
-    {{ form.csrf }}
+  <form method="POST">
     <div class="form-group">
-      {% for field in form %}
       <div class="form-field">
-        <div class="label-field">{{ field.label }}</div>
-        <div class="input-field">{{ field }}</div>
+        <div class="label-field">Enter your username</div>
+        <input name="username">
       </div>
-      {% endfor %}
+      <div class="form-field">
+        <div class="label-field">Password</div>
+        <input type="password" name="password">
+      </div>
+      <div class="form-field">
+        <div class="label-field">Confirm password</div>
+        <input type="password" name="confirm-password">
+      </div>
+      <div class="form-field">
+        <div class="label-field">Enter address (optional)</div>
+        <input name="email">
+      </div>
+      <div class="form-field">
+        <div class="label-field">Enter the text you see on the image below</div>
+        <input name="captcha-word">
+      </div>
       <img src="{{ captcha }}">
     </div>
     <button type="submit" class="btn btn-success">Create your account</button>
   </form>
+
   <br>
   {% with messages = get_flashed_messages(with_categories=true) %}
   {% if messages %}
   {% for message in messages %}
-  {% if "Error" not in message[1]: %}
   <div class="alert alert-info">
     {{ message[1] }}
   </div>
-  {% endif %}
-  {% if "Error" in message[1]: %}
-  <div class="alert alert-warning">
-    {{ message[1] }}
-  </div>
-  {% endif %}
   {% endfor %}
   {% endif %}
   {% endwith %}
